Validações, filtering e sorting

+++ Controllers/CommentController.cs
@@ -22,6 +22,9 @@ namespace RESTAPI.Controllers
 
         public async Task<IActionResult> GetAll()
         {
+
+            if(!ModelState.IsValid) { return BadRequest(); }
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -35,6 +38,9 @@ namespace RESTAPI.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> GetById(int id)
         {
+
+            if (!ModelState.IsValid) { return BadRequest(); }
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if (comment == null) { return NotFound(); };
@@ -48,6 +54,8 @@ namespace RESTAPI.Controllers
 
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentRequestDto commentDto)
         {
+            if (!ModelState.IsValid) { return BadRequest(); }
+
 
             if (!await _stockRepo.StockExists(stockId))
             {
@@ -71,6 +79,9 @@ namespace RESTAPI.Controllers
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
 
+            if (!ModelState.IsValid) { return BadRequest(); }
+
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if(comment == null)
@@ -85,6 +96,9 @@ namespace RESTAPI.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Delete ([FromRoute] int id)
         {
+
+            if (!ModelState.IsValid) { return BadRequest(); }
+
             var comment = await _commentRepo.DeleteAsync(id);
 
             if(comment == null) { return NotFound("Comment Does not exist"); }
